Change int to double in driveDistance method and related tests, all passing

+++ behaviours/Driveable.java
@@ -1,5 +1,5 @@
 package behaviours;
 
 public interface Driveable {
-  Integer driveDistance(int distance);
+  double driveDistance(double distance);
 }
